Addressed some iOS 7.1 build issues and 64-bit warnings

I'll be changing the pbxproj's reference to libz and Foundation in the iPhone SDK so it's not version-specific.

+++ ZipKit/NSData+ZKAdditions.h
@@ -14,8 +14,8 @@
 - (UInt64) zk_hostInt64OffsetBy:(UInt64 *)offset;
 - (BOOL) zk_hostBoolOffsetBy:(UInt64 *)offset;
 - (NSString *) zk_stringOffsetBy:(UInt64 *)offset length:(NSUInteger)length;
-- (unsigned long) zk_crc32;
-- (unsigned long) zk_crc32:(unsigned long)crc;
+- (UInt32) zk_crc32;
+- (UInt32) zk_crc32:(unsigned long)crc;
 
 - (NSData *) zk_inflateWithWindowBits:(int)windowBits;
 - (NSData *) zk_deflateWithLevel:(int)level windowBits:(int)windowBits memoryLevel:(int)memoryLevel strategy:(int)strategy;
