+++ ZipKit/ZKCDTrailer64.m
@@ -25,7 +25,7 @@
 
 + (ZKCDTrailer64 *) recordWithData:(NSData *)data atOffset:(UInt64)offset {
 	if (!data) return nil;
-	NSUInteger mn = [data zk_hostInt32OffsetBy:&offset];
+	UInt32 mn = [data zk_hostInt32OffsetBy:&offset];
 	if (mn != ZKCDTrailer64MagicNumber) return nil;
 	ZKCDTrailer64 *record = [ZKCDTrailer64 new];
 	record.magicNumber = mn;
